Refactor inventory management to support decimal quantities
- Updated inventory adjustment dialog to handle double values for quantities instead of integers.
- Modified procurement and product screens to use double for stock and price values.
- Refactored API service calls to accommodate changes in data types.
- Removed deprecated secure storage references and updated dependencies.
- Enhanced validation for quantity inputs in various forms to allow decimal values.
- Cleaned up code by removing unused imports and consolidating API service usage.

+++ lib/features/admin/procurement/screens/admin_procurement_create_screen.dart
@@ -213,7 +213,7 @@ class _AdminProcurementCreateScreenState
     required String productId,
     required String productName,
     required double defaultUnitCost,
-    required int currentStock,
+    required double currentStock,
   }) async {
     final existing = _cart[productId];
 
@@ -376,8 +376,8 @@ class _PickStoreHint extends StatelessWidget {
 class _ProductCard extends StatelessWidget {
   final String name;
   final double price;
-  final int stock;
-  final int minStock;
+  final double stock;
+  final double minStock;
   final bool selected;
   final int selectedQty;
   final VoidCallback? onTap;
@@ -503,7 +503,7 @@ class _AddItemSheet extends StatefulWidget {
   final String productName;
   final int initialQty;
   final double initialUnitCost;
-  final int maxStock;
+  final double maxStock;
   final VoidCallback? onRemove;
 
   const _AddItemSheet({
